refactor: 로깅 패키지 이동
