+++ src/main/java/back/ecommerce/common/logging/GlobalLogger.java
@@ -1,4 +1,4 @@
-package back.ecommerce.common;
+package back.ecommerce.common.logging;
 
 import java.util.EnumMap;
