+++ src/main/java/back/ecommerce/common/logging/SearchRequestLogger.java
@@ -1,4 +1,4 @@
-package back.ecommerce.common;
+package back.ecommerce.common.logging;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
