+++ src/main/java/back/ecommerce/config/logger/LoggerConfig.java
@@ -3,7 +3,7 @@ package back.ecommerce.config.logger;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import back.ecommerce.common.GlobalLogger;
+import back.ecommerce.common.logging.GlobalLogger;
 
 @Configuration
 public class LoggerConfig {
