+++ src/test/java/back/ecommerce/controller/CartControllerTest.java
@@ -34,8 +34,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import back.ecommerce.auth.annotaion.UserEmail;
 import back.ecommerce.controller.cart.CartController;
-import back.ecommerce.common.GlobalExceptionHandler;
-import back.ecommerce.common.GlobalLogger;
+import back.ecommerce.common.logging.GlobalExceptionHandler;
+import back.ecommerce.common.logging.GlobalLogger;
 import back.ecommerce.domain.product.Category;
 import back.ecommerce.dto.request.cart.AddCartRequest;
 import back.ecommerce.dto.response.cart.AddCartResponse;
